rename panel to quality-menu-injector
