+++ tubesize/src/quality-menu-injector.ts
